extract information from retained fields in Claim component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Claim component of FHIR Bundle, produce simple data.frame

process_Claim(Claim)

Arguments

Claim

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Claim(tbun$Claim)
#>                                       id
#> 7   f7259a2c-0399-45bb-985f-094f4fe4c021
#> 18  f4b485ea-2527-49c1-8f88-c3941b40ec6f
#> 23  bd80500e-4dab-4022-8503-aaf3c290ee36
#> 24  b342e379-84aa-4076-9cc4-1206bd6cb6ff
#> 28  b8b23adb-1949-483a-95f8-93f3b93d1768
#> 53  62cbf8c1-586c-489b-b4c2-ae962759c149
#> 64  58f72e23-45f4-4c18-943b-2d5d76ca99d6
#> 75  2875b6bd-da44-49c2-8cb2-238d3d49406a
#> 79  e785a3ab-73fd-4fc3-bdc6-a324a69a15b6
#> 83  ecf41b47-280b-4a7a-b8f2-f2700a6727a4
#> 89  82b5d6c2-c993-43a6-a7ea-521439e4852e
#> 90  5c319f1a-7f31-4b09-891d-6574a9f3be6d
#> 93  7f1335a7-9c13-4d4c-8325-ffa5c6b5d2a7
#> 104 e884a2c2-2607-42ff-a120-77660754a877
#> 109 12aba208-bf65-4ba4-ad29-794c292c8ba1
#> 110 96888531-88db-47a6-8a77-7e8e9fcefe70
#> 113 75bebb3c-c368-49d9-bab6-e51ecf7e979f
#> 119 73c1f551-0a53-4395-af17-b4c92c68aa62
#> 122 fc46105f-b3b7-44f9-81aa-abe8c753bf70
#> 125 ec3c0b45-55b1-469c-8f89-67f720450dbc
#> 136 0452a7b1-7acc-4d3e-918c-6d7a945904e7
#> 139 f83cd246-4d21-46df-88d3-e13205eeac0f
#> 149 bcdf1322-bf6c-4bae-bfbd-40679ff5dcff
#> 154 d73e2f67-9e32-4a5c-90e7-4da824fca00b
#> 157 44fef3fe-df88-447a-a817-7aaa90e78ea6
#> 160 200c62ca-f6d1-40aa-b1fb-89de454d7957
#> 164 1e5d2ea3-6b3b-4c19-a9f4-07bf691d2944
#> 168 3dd7cee1-43a4-4aed-b1d4-1c8c45cbb562
#> 192 3c79711d-066c-4e60-a528-59b6c310f886
#> 203 2d1ffa67-3ba1-4cee-af8b-f69a1b8c2b13
#> 208 c40b1fbe-a2e9-4231-9e43-6791f8cdbe7c
#> 211 986fb60f-5f8a-466f-bd72-0ac12ef51efe
#> 217 1514d31b-fd6b-4790-adc0-3ec6aede7a2f
#> 218 f0994621-4211-4f99-b200-e36bc7eb8a23
#> 221 83637777-ee3a-478a-9841-be357ccb6f2f
#> 232 930f1e2c-50a0-4187-b86e-d6234e0a9829
#> 235 773334a1-6e9f-4c1a-a350-d4e14dab72ea
#> 247 614b7593-0241-4e75-a265-52c3cc5a8eea
#> 255 5792a5e1-90ce-44e0-a628-f1614e7265a0
#> 256 a18ce70d-5dae-4113-bc1a-ebe37bef522f
#> 268 9c23b9b7-7452-4348-a8f6-8cdd8de0a09d
#> 280 ef1e7d54-3798-4e48-82df-ac06089d760a
#> 282 6c5f6f70-fcf7-4233-a6a9-f98af7855bc4
#> 293 c1cbaf01-8e7f-4bbc-a6d7-82c9fa2001d7
#> 295 f0b4b732-490f-4dba-8c40-cad5f5962185
#> 300 ba0a2b8e-da8d-41ca-9243-2fa1e26560b1
#>                                 patient.reference   patient.display
#> 7   urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 18  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 23  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 24  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 28  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 53  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 64  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 75  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 79  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 83  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 89  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 90  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 93  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 104 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 109 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 110 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 113 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 119 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 122 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 125 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 136 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 139 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 149 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 154 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 157 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 160 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 164 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 168 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 192 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 203 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 208 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 211 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 217 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 218 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 221 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 232 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 235 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 247 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 255 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 256 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 268 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 280 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 282 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 293 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2              <NA>
#> 295 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#> 300 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 Vince741 Rogahn59
#>                                provider.reference
#> 7   urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 18  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 23  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 24  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 28  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 53  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 64  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 75  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 79  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 83  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 89  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 90  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 93  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 104 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 109 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 110 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 113 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 119 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 122 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 125 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 136 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 139 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 149 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 154 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 157 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 160 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 164 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 168 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 192 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 203 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 208 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 211 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 217 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 218 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 221 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 232 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 235 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 247 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 255 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 256 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 268 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 280 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 282 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 293 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 295 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 300 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#>                           provider.display insurance.sequence insurance.focal
#> 7                   HALLMARK HEALTH SYSTEM                  1            TRUE
#> 18                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 23                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 24                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 28                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 53  HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 64  HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 75  HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 79                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 83                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 89                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 90                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 93                  HALLMARK HEALTH SYSTEM                  1            TRUE
#> 104 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 109                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 110                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 113                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 119             PROMPT CARE WALK-IN CLINIC                  1            TRUE
#> 122                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 125             PROMPT CARE WALK-IN CLINIC                  1            TRUE
#> 136 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 139                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 149                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 154                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 157                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 160             PROMPT CARE WALK-IN CLINIC                  1            TRUE
#> 164                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 168                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 192                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 203 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 208                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 211                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 217                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 218                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 221                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 232 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 235                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 247                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 255                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 256                 HALLMARK HEALTH SYSTEM                  1            TRUE
#> 268 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 280 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 282 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 293 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 295 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#> 300 HALLMARK HEALTH MEDICAL ASSOCIATES INC                  1            TRUE
#>          display      billablePeriod.start        billablePeriod.end
#> 7   NO_INSURANCE 1972-06-17T07:25:09-04:00 1972-06-17T07:40:09-04:00
#> 18  NO_INSURANCE 1972-06-30T07:25:09-04:00 1972-06-30T07:40:09-04:00
#> 23        Anthem 1973-06-29T07:25:09-04:00 1973-07-13T07:25:09-04:00
#> 24  NO_INSURANCE 1973-06-29T07:25:09-04:00 1973-07-13T07:25:09-04:00
#> 28  NO_INSURANCE 1977-03-29T06:25:09-05:00 1977-03-29T06:40:09-05:00
#> 53        Anthem 1977-06-08T07:25:09-04:00 1977-06-08T07:40:09-04:00
#> 64        Anthem 1978-06-14T07:25:09-04:00 1978-06-14T07:40:09-04:00
#> 75        Anthem 1979-06-20T07:25:09-04:00 1979-06-20T07:40:09-04:00
#> 79  NO_INSURANCE 1979-07-19T07:25:09-04:00 1979-07-19T07:40:09-04:00
#> 83  NO_INSURANCE 1979-08-16T07:25:09-04:00 1979-08-16T07:40:09-04:00
#> 89        Anthem 1980-01-07T06:25:09-05:00 1980-01-07T09:40:09-05:00
#> 90  NO_INSURANCE 1980-01-07T06:25:09-05:00 1980-01-07T09:40:09-05:00
#> 93  NO_INSURANCE 1980-01-09T06:25:09-05:00 1980-01-09T06:40:09-05:00
#> 104       Anthem 1980-06-25T07:25:09-04:00 1980-06-25T07:40:09-04:00
#> 109       Anthem 1981-03-14T06:25:09-05:00 1981-03-14T06:40:09-05:00
#> 110 NO_INSURANCE 1981-03-14T06:25:09-05:00 1981-03-14T06:40:09-05:00
#> 113 NO_INSURANCE 1981-05-16T07:25:09-04:00 1981-05-16T07:40:09-04:00
#> 119       Anthem 1981-05-27T07:25:09-04:00 1981-05-27T07:40:09-04:00
#> 122 NO_INSURANCE 1981-05-30T07:25:09-04:00 1981-05-30T07:40:09-04:00
#> 125       Anthem 1981-06-10T07:25:09-04:00 1981-06-10T07:40:09-04:00
#> 136       Anthem 1981-07-01T07:25:09-04:00 1981-07-01T07:55:09-04:00
#> 139 NO_INSURANCE 1981-07-04T07:25:09-04:00 1981-07-04T07:40:09-04:00
#> 149       Anthem 1981-07-15T07:25:09-04:00 1981-07-15T07:40:09-04:00
#> 154 NO_INSURANCE 1981-09-28T07:25:09-04:00 1981-09-28T08:25:09-04:00
#> 157 NO_INSURANCE 1981-09-28T07:25:09-04:00 1981-09-28T08:25:09-04:00
#> 160       Anthem 1981-11-18T06:25:09-05:00 1981-11-18T06:40:09-05:00
#> 164 NO_INSURANCE 1981-11-17T06:25:09-05:00 1981-11-17T06:40:09-05:00
#> 168       Anthem 1982-06-16T07:25:09-04:00 1982-06-16T07:40:09-04:00
#> 192 NO_INSURANCE 1982-06-12T07:25:09-04:00 1982-06-12T07:40:09-04:00
#> 203       Anthem 1982-07-07T07:25:09-04:00 1982-07-07T07:55:09-04:00
#> 208 NO_INSURANCE 1982-11-13T06:25:09-05:00 1982-11-13T07:25:09-05:00
#> 211 NO_INSURANCE 1982-11-13T06:25:09-05:00 1982-11-13T07:25:09-05:00
#> 217       Anthem 1983-03-24T06:25:09-05:00 1983-03-24T09:40:09-05:00
#> 218 NO_INSURANCE 1983-03-24T06:25:09-05:00 1983-03-24T09:40:09-05:00
#> 221 NO_INSURANCE 1983-03-30T06:25:09-05:00 1983-03-30T06:40:09-05:00
#> 232       Anthem 1983-07-13T07:25:09-04:00 1983-07-13T07:40:09-04:00
#> 235       Anthem 1983-08-31T07:25:09-04:00 1983-08-31T07:40:09-04:00
#> 247 NO_INSURANCE 1983-08-25T07:25:09-04:00 1983-08-25T07:40:09-04:00
#> 255 NO_INSURANCE 1984-06-19T07:25:09-04:00 1984-06-19T11:10:09-04:00
#> 256 NO_INSURANCE 1984-06-19T07:25:09-04:00 1984-06-19T11:10:09-04:00
#> 268       Anthem 1984-07-18T07:25:09-04:00 1984-07-18T07:55:09-04:00
#> 280 NO_INSURANCE 1985-07-24T07:25:09-04:00 1985-07-24T07:55:09-04:00
#> 282       Anthem 1985-07-24T07:25:09-04:00 1985-07-24T07:55:09-04:00
#> 293       Anthem 1986-07-30T07:25:09-04:00 1986-07-30T07:40:09-04:00
#> 295       Anthem 1986-07-30T07:25:09-04:00 1986-07-30T07:40:09-04:00
#> 300       Anthem 1987-05-20T07:25:09-04:00 1987-05-20T07:40:09-04:00
#>                       created
#> 7   1972-06-17T07:40:09-04:00
#> 18  1972-06-30T07:40:09-04:00
#> 23  1973-07-13T07:25:09-04:00
#> 24  1973-07-13T07:25:09-04:00
#> 28  1977-03-29T06:40:09-05:00
#> 53  1977-06-08T07:40:09-04:00
#> 64  1978-06-14T07:40:09-04:00
#> 75  1979-06-20T07:40:09-04:00
#> 79  1979-07-19T07:40:09-04:00
#> 83  1979-08-16T07:40:09-04:00
#> 89  1980-01-07T09:40:09-05:00
#> 90  1980-01-07T09:40:09-05:00
#> 93  1980-01-09T06:40:09-05:00
#> 104 1980-06-25T07:40:09-04:00
#> 109 1981-03-14T06:40:09-05:00
#> 110 1981-03-14T06:40:09-05:00
#> 113 1981-05-16T07:40:09-04:00
#> 119 1981-05-27T07:40:09-04:00
#> 122 1981-05-30T07:40:09-04:00
#> 125 1981-06-10T07:40:09-04:00
#> 136 1981-07-01T07:55:09-04:00
#> 139 1981-07-04T07:40:09-04:00
#> 149 1981-07-15T07:40:09-04:00
#> 154 1981-09-28T08:25:09-04:00
#> 157 1981-09-28T08:25:09-04:00
#> 160 1981-11-18T06:40:09-05:00
#> 164 1981-11-17T06:40:09-05:00
#> 168 1982-06-16T07:40:09-04:00
#> 192 1982-06-12T07:40:09-04:00
#> 203 1982-07-07T07:55:09-04:00
#> 208 1982-11-13T07:25:09-05:00
#> 211 1982-11-13T07:25:09-05:00
#> 217 1983-03-24T09:40:09-05:00
#> 218 1983-03-24T09:40:09-05:00
#> 221 1983-03-30T06:40:09-05:00
#> 232 1983-07-13T07:40:09-04:00
#> 235 1983-08-31T07:40:09-04:00
#> 247 1983-08-25T07:40:09-04:00
#> 255 1984-06-19T11:10:09-04:00
#> 256 1984-06-19T11:10:09-04:00
#> 268 1984-07-18T07:55:09-04:00
#> 280 1985-07-24T07:55:09-04:00
#> 282 1985-07-24T07:55:09-04:00
#> 293 1986-07-30T07:40:09-04:00
#> 295 1986-07-30T07:40:09-04:00
#> 300 1987-05-20T07:40:09-04:00