extract information from retained fields in Condition component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Condition component of FHIR Bundle, produce simple data.frame

process_Condition(Condition)

Arguments

Condition

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Condition(tbun$Condition)
#>                                      id
#> 1  adc05f8e-0e90-4217-b1d7-fa41b8e33638
#> 2  a7c0f1bf-f590-4264-9d23-7d572814097b
#> 3  793a03ca-7c3e-4410-99ca-b8b88651610b
#> 4  bdd14588-4969-45b2-9170-e4ae6ef9f8ca
#> 5  d3c797f5-b972-4047-b829-1ec32084e1e1
#> 6  91b48a2b-d649-4ee9-8dc8-170956480df6
#> 7  66e90602-4775-45cd-b7f0-4d7655658020
#> 8  ed286f51-d8b8-4cdd-bfd3-3a49b96026ec
#> 9  118eecc5-8594-4c88-a5d1-eae9d6975b7a
#> 10 98d3e627-4e0c-414b-8b33-070d6c13b17d
#> 11 a26ddcf6-c4dd-47a9-b3c1-5a4fc48f83a7
#> 12 db492aea-c5ce-4421-bdab-0c4172f5f341
#> 13 1c027d75-8838-47bf-98ea-36e05478d6bf
#> 14 b7a2b401-4e75-461a-a5ce-24737861c145
#> 15 a7e1e1ba-240c-46ce-aa4f-83e3ee7f9cd2
#>                                subject.reference     code.coding.system
#> 1  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 2  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 3  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 4  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 5  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 6  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 7  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 8  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 9  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 10 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 11 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 12 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 13 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 14 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#> 15 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://snomed.info/sct
#>    code.coding.code                   code.coding.display
#> 1         446096008           Perennial allergic rhinitis
#> 2          65363002                          Otitis media
#> 3          75498004  Acute bacterial sinusitis (disorder)
#> 4          40055000          Chronic sinusitis (disorder)
#> 5         241929008               Acute allergic reaction
#> 6          36971009                  Sinusitis (disorder)
#> 7          44465007                       Sprain of ankle
#> 8         444814009            Viral sinusitis (disorder)
#> 9         444814009            Viral sinusitis (disorder)
#> 10         44465007                       Sprain of ankle
#> 11        241929008               Acute allergic reaction
#> 12        444814009            Viral sinusitis (disorder)
#> 13        128613002                      Seizure disorder
#> 14        703151001 History of single seizure (situation)
#> 15         84757009                              Epilepsy