extract information from retained fields in Encounter component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Encounter component of FHIR Bundle, produce simple data.frame

process_Encounter(Encounter)

Arguments

Encounter

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Encounter(tbun$Encounter)
#>                                       id            type.system type.code
#> 4   a2dbf45f-4d45-4d15-9519-eea32b9b1d44 http://snomed.info/sct 185347001
#> 9   b98c91d5-53ee-4fb9-bf49-2f92982e0453 http://snomed.info/sct 185347001
#> 20  9c95930d-83a4-4a89-9bbb-8e540e729233 http://snomed.info/sct 185345009
#> 26  d9e4bd61-2d1a-4af5-a953-32daa700daed http://snomed.info/sct 185345009
#> 32  6e14a928-a451-4de0-b2de-8ca493ecffa8 http://snomed.info/sct 410620009
#> 55  4294bcd8-8929-41a4-9c77-3cc3ec3d5325 http://snomed.info/sct 410620009
#> 66  084a8e7f-ebe1-4985-bf6a-2eb566d46b61 http://snomed.info/sct 410620009
#> 77  7a68e19f-4be3-42c6-aab0-8ec86dc3e9df http://snomed.info/sct 185345009
#> 81  95d30a72-3155-4206-8b59-a19e313d922a http://snomed.info/sct 185345009
#> 85  1a867cb1-ae9a-4175-af1c-9bc1d6b5a5cb http://snomed.info/sct  50849002
#> 92  55d3e698-3af8-4d71-b3f8-3efe405221ef http://snomed.info/sct 185347001
#> 95  5dc3e521-7bd3-4820-9582-7cbad84bf341 http://snomed.info/sct 410620009
#> 106 ec6accdf-00b7-456c-afbe-b26fc65d018f http://snomed.info/sct 185345009
#> 112 352eb560-51aa-48dd-861e-9e1abad79fc6 http://snomed.info/sct 185345009
#> 117 039d6893-17a7-4b06-8ecf-0bd7e6db873c http://snomed.info/sct 702927004
#> 121 358b11fc-85f2-4755-9f2f-b562252f6c46 http://snomed.info/sct 185345009
#> 124 33f9e2ef-4f3b-411d-8df3-24ce3a79a807 http://snomed.info/sct 702927004
#> 127 f79e752f-db7c-45bc-8c2f-bac7d4cb77c8 http://snomed.info/sct 410620009
#> 138 45bb97d8-06ca-44c6-b0be-a648be778f87 http://snomed.info/sct 185345009
#> 141 0e1f8016-6ff4-461a-9dc3-fdab472e4343 http://snomed.info/sct 185349003
#> 151 dafb437b-17d5-45d6-8e26-b95aa6f9c452 http://snomed.info/sct  50849002
#> 159 aa7e17fb-9add-4552-9c98-cd1db3c836d6 http://snomed.info/sct 702927004
#> 162 da51a412-6934-4c09-bbfe-858922c8b4ab http://snomed.info/sct 185345009
#> 166 37c36d3e-8948-4901-b877-ae0670e084dc http://snomed.info/sct 185349003
#> 170 c31b5913-7842-4975-b81f-18c66dd0fe23 http://snomed.info/sct 185345009
#> 194 285e30f3-5dec-41ed-99f1-a0bfa9369e24 http://snomed.info/sct 410620009
#> 205 f0ce3e3e-4d55-45e4-b54b-2093fa9f3a0f http://snomed.info/sct  50849002
#> 213 77d50a55-6dc7-4f4a-9c40-ee90a2906f48 http://snomed.info/sct  50849002
#> 220 1b9e47a9-f4df-4d51-ba2d-b59b16c7dc82 http://snomed.info/sct 185347001
#> 223 b30f53e2-d4c7-40ef-9446-23c8a94dd699 http://snomed.info/sct 410620009
#> 234 b8a97ee6-32dc-43c4-ba7e-1791045741e9 http://snomed.info/sct 185349003
#> 237 d24e92ec-86d5-4462-a02b-7c9988464d5a http://snomed.info/sct 185345009
#> 249 23ab1996-cb11-4ca6-bad1-a1a0a1357f22 http://snomed.info/sct  50849002
#> 258 bc2c0896-996f-4d0d-a452-6fd03de2f59c http://snomed.info/sct 410620009
#> 270 83463504-3f06-4c10-abe9-4dc615427562 http://snomed.info/sct 410620009
#> 284 258e175c-db20-447d-addc-927a1aea2ddc http://snomed.info/sct 410620009
#> 297 88d40336-c72e-4a4e-9c30-734e298cdce7 http://snomed.info/sct 308646001
#>                             type.display
#> 4                  Encounter for problem
#> 9                  Encounter for problem
#> 20                 Encounter for symptom
#> 26                 Encounter for symptom
#> 32          Well child visit (procedure)
#> 55          Well child visit (procedure)
#> 66          Well child visit (procedure)
#> 77                 Encounter for symptom
#> 81                 Encounter for symptom
#> 85  Emergency room admission (procedure)
#> 92                 Encounter for problem
#> 95          Well child visit (procedure)
#> 106                Encounter for symptom
#> 112                Encounter for symptom
#> 117       Urgent care clinic (procedure)
#> 121                Encounter for symptom
#> 124       Urgent care clinic (procedure)
#> 127         Well child visit (procedure)
#> 138                Encounter for symptom
#> 141   Encounter for check up (procedure)
#> 151 Emergency room admission (procedure)
#> 159       Urgent care clinic (procedure)
#> 162                Encounter for symptom
#> 166   Encounter for check up (procedure)
#> 170                Encounter for symptom
#> 194         Well child visit (procedure)
#> 205 Emergency room admission (procedure)
#> 213 Emergency room admission (procedure)
#> 220                Encounter for problem
#> 223         Well child visit (procedure)
#> 234   Encounter for check up (procedure)
#> 237                Encounter for symptom
#> 249             Emergency Room Admission
#> 258         Well child visit (procedure)
#> 270         Well child visit (procedure)
#> 284         Well child visit (procedure)
#> 297                  Death Certification
#>                                           subject
#> 4   urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 9   urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 20  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 26  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 32  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 55  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 66  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 77  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 81  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 85  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 92  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 95  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 106 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 112 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 117 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 121 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 124 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 127 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 138 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 141 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 151 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 159 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 162 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 166 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 170 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 194 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 205 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 213 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 220 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 223 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 234 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 237 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 249 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 258 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 270 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 284 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#> 297 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2
#>                         serviceProvider.reference
#> 4   urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 9   urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 20  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 26  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 32  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 55  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 66  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 77  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 81  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 85  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 92  urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 95  urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 106 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 112 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 117 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 121 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 124 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 127 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 138 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 141 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 151 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 159 urn:uuid:90cf148c-69ed-3d33-aa86-b509c6b0b25f
#> 162 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 166 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 170 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 194 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 205 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 213 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 220 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 223 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 234 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 237 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 249 urn:uuid:d692e283-0833-3201-8e55-4f868a9c0736
#> 258 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 270 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 284 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#> 297 urn:uuid:6a6909c5-55ad-35e4-8f1c-7b0407c5b691
#>                    serviceProvider.display              period.start
#> 4                   HALLMARK HEALTH SYSTEM 1972-06-17T07:25:09-04:00
#> 9                   HALLMARK HEALTH SYSTEM 1972-06-30T07:25:09-04:00
#> 20                  HALLMARK HEALTH SYSTEM 1973-06-29T07:25:09-04:00
#> 26                  HALLMARK HEALTH SYSTEM 1977-03-29T06:25:09-05:00
#> 32  HALLMARK HEALTH MEDICAL ASSOCIATES INC 1977-06-08T07:25:09-04:00
#> 55  HALLMARK HEALTH MEDICAL ASSOCIATES INC 1978-06-14T07:25:09-04:00
#> 66  HALLMARK HEALTH MEDICAL ASSOCIATES INC 1979-06-20T07:25:09-04:00
#> 77                  HALLMARK HEALTH SYSTEM 1979-07-19T07:25:09-04:00
#> 81                  HALLMARK HEALTH SYSTEM 1979-08-16T07:25:09-04:00
#> 85                  HALLMARK HEALTH SYSTEM 1980-01-07T06:25:09-05:00
#> 92                  HALLMARK HEALTH SYSTEM 1980-01-09T06:25:09-05:00
#> 95  HALLMARK HEALTH MEDICAL ASSOCIATES INC 1980-06-25T07:25:09-04:00
#> 106                 HALLMARK HEALTH SYSTEM 1981-03-14T06:25:09-05:00
#> 112                 HALLMARK HEALTH SYSTEM 1981-05-16T07:25:09-04:00
#> 117             PROMPT CARE WALK-IN CLINIC 1981-05-27T07:25:09-04:00
#> 121                 HALLMARK HEALTH SYSTEM 1981-05-30T07:25:09-04:00
#> 124             PROMPT CARE WALK-IN CLINIC 1981-06-10T07:25:09-04:00
#> 127 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1981-07-01T07:25:09-04:00
#> 138                 HALLMARK HEALTH SYSTEM 1981-07-04T07:25:09-04:00
#> 141                 HALLMARK HEALTH SYSTEM 1981-07-15T07:25:09-04:00
#> 151                 HALLMARK HEALTH SYSTEM 1981-09-28T07:25:09-04:00
#> 159             PROMPT CARE WALK-IN CLINIC 1981-11-18T06:25:09-05:00
#> 162                 HALLMARK HEALTH SYSTEM 1981-11-17T06:25:09-05:00
#> 166                 HALLMARK HEALTH SYSTEM 1982-06-16T07:25:09-04:00
#> 170                 HALLMARK HEALTH SYSTEM 1982-06-12T07:25:09-04:00
#> 194 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1982-07-07T07:25:09-04:00
#> 205                 HALLMARK HEALTH SYSTEM 1982-11-13T06:25:09-05:00
#> 213                 HALLMARK HEALTH SYSTEM 1983-03-24T06:25:09-05:00
#> 220                 HALLMARK HEALTH SYSTEM 1983-03-30T06:25:09-05:00
#> 223 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1983-07-13T07:25:09-04:00
#> 234                 HALLMARK HEALTH SYSTEM 1983-08-31T07:25:09-04:00
#> 237                 HALLMARK HEALTH SYSTEM 1983-08-25T07:25:09-04:00
#> 249                 HALLMARK HEALTH SYSTEM 1984-06-19T07:25:09-04:00
#> 258 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1984-07-18T07:25:09-04:00
#> 270 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1985-07-24T07:25:09-04:00
#> 284 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1986-07-30T07:25:09-04:00
#> 297 HALLMARK HEALTH MEDICAL ASSOCIATES INC 1987-05-20T07:25:09-04:00
#>                    period.end
#> 4   1972-06-17T07:40:09-04:00
#> 9   1972-06-30T07:40:09-04:00
#> 20  1973-07-13T07:25:09-04:00
#> 26  1977-03-29T06:40:09-05:00
#> 32  1977-06-08T07:40:09-04:00
#> 55  1978-06-14T07:40:09-04:00
#> 66  1979-06-20T07:40:09-04:00
#> 77  1979-07-19T07:40:09-04:00
#> 81  1979-08-16T07:40:09-04:00
#> 85  1980-01-07T09:40:09-05:00
#> 92  1980-01-09T06:40:09-05:00
#> 95  1980-06-25T07:40:09-04:00
#> 106 1981-03-14T06:40:09-05:00
#> 112 1981-05-16T07:40:09-04:00
#> 117 1981-05-27T07:40:09-04:00
#> 121 1981-05-30T07:40:09-04:00
#> 124 1981-06-10T07:40:09-04:00
#> 127 1981-07-01T07:55:09-04:00
#> 138 1981-07-04T07:40:09-04:00
#> 141 1981-07-15T07:40:09-04:00
#> 151 1981-09-28T08:25:09-04:00
#> 159 1981-11-18T06:40:09-05:00
#> 162 1981-11-17T06:40:09-05:00
#> 166 1982-06-16T07:40:09-04:00
#> 170 1982-06-12T07:40:09-04:00
#> 194 1982-07-07T07:55:09-04:00
#> 205 1982-11-13T07:25:09-05:00
#> 213 1983-03-24T09:40:09-05:00
#> 220 1983-03-30T06:40:09-05:00
#> 223 1983-07-13T07:40:09-04:00
#> 234 1983-08-31T07:40:09-04:00
#> 237 1983-08-25T07:40:09-04:00
#> 249 1984-06-19T11:10:09-04:00
#> 258 1984-07-18T07:55:09-04:00
#> 270 1985-07-24T07:55:09-04:00
#> 284 1986-07-30T07:40:09-04:00
#> 297 1987-05-20T07:40:09-04:00