extract information from retained fields in Immunization component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Immunization component of FHIR Bundle, produce simple data.frame

process_Immunization(Immunization)

Arguments

Immunization

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Immunization(tbun$Immunization)
#>                                      id
#> 1  b252b8a6-0560-4f4e-864a-a9f839c453ec
#> 2  7c2ad2e8-7044-47a2-9da8-248137b2fcf7
#> 3  82b58916-19dd-4be6-970b-ddb4a32a44e3
#> 4  24efb1ba-da41-4f3b-953d-acabaa75cb35
#> 5  8820744c-0984-410f-a847-b6e5476e765b
#> 6  7fc4e7d6-877b-40c7-8448-dcfc1c81773d
#> 7  21c25847-66f3-415a-b9ec-5f29c8676635
#> 8  3f2cdb65-556b-4055-99be-5aa45df153d7
#> 9  e47c0ae5-23c9-47ff-be0d-777116e9410a
#> 10 06b5c37f-c7bd-422f-a29f-a70d1d95ebd6
#>                                          patient          vaccineCode.system
#> 1  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 2  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 3  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 4  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 5  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 6  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 7  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 8  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 9  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#> 10 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 http://hl7.org/fhir/sid/cvx
#>    vaccineCode.code                                vaccineCode.display
#> 1               140 Influenza, seasonal, injectable, preservative free
#> 2               140 Influenza, seasonal, injectable, preservative free
#> 3               140 Influenza, seasonal, injectable, preservative free
#> 4               140 Influenza, seasonal, injectable, preservative free
#> 5               140 Influenza, seasonal, injectable, preservative free
#> 6               140 Influenza, seasonal, injectable, preservative free
#> 7               140 Influenza, seasonal, injectable, preservative free
#> 8               140 Influenza, seasonal, injectable, preservative free
#> 9               140 Influenza, seasonal, injectable, preservative free
#> 10              140 Influenza, seasonal, injectable, preservative free
#>           occurrenceDateTime
#> 1  1977-06-08T07:25:09-04:00
#> 2  1978-06-14T07:25:09-04:00
#> 3  1979-06-20T07:25:09-04:00
#> 4  1980-06-25T07:25:09-04:00
#> 5  1981-05-27T07:25:09-04:00
#> 6  1982-06-16T07:25:09-04:00
#> 7  1983-07-13T07:25:09-04:00
#> 8  1984-07-18T07:25:09-04:00
#> 9  1985-07-24T07:25:09-04:00
#> 10 1986-07-30T07:25:09-04:00