extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame

process_Observation(Observation)

Arguments

Observation

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Observation(tbun$Observation)
#>                                       id code.coding.system code.coding.code
#> 1   d8d29164-6f51-411f-9ac3-a8fd4746bbb4   http://loinc.org           8302-2
#> 2   0cb0cc06-d732-45cf-b79a-29127915baa2   http://loinc.org          72514-3
#> 3   e16a09b0-b01f-44cb-b301-d606729e4b38   http://loinc.org          29463-7
#> 4   9625b98f-bc58-453c-8664-f1fe185840a8   http://loinc.org          39156-5
#> 5   690c8151-8c6b-4c1a-81d9-1071818592cc   http://loinc.org          59576-9
#> 6   a91b50ea-2596-4f2c-9a3c-4d38054a89eb   http://loinc.org          55284-4
#> 7   74b16bae-d8b1-4cf9-9281-ad3736348641   http://loinc.org           6690-2
#> 8   911c76fe-c953-4bf2-ba74-d71b44399296   http://loinc.org            789-8
#> 9   013b4774-cea5-437b-a6ea-1e0e0dcde8d9   http://loinc.org            718-7
#> 10  895ccd65-5bec-4963-9bfd-824865121280   http://loinc.org           4544-3
#> 11  0d72572d-2c30-4d7d-8aa2-14475e2f3962   http://loinc.org            787-2
#> 12  810cc728-d102-44c1-9ea0-6b08f2d0edc5   http://loinc.org            785-6
#> 13  899058cb-26ec-4be6-bd1c-43cac0356d54   http://loinc.org            786-4
#> 14  9ef4647c-d144-4195-a8c4-89ab15910d1b   http://loinc.org          21000-5
#> 15  b4b17f38-8920-4634-a9ce-b0d2e7833bcc   http://loinc.org            777-3
#> 16  18b51ef9-e7d7-4656-95f8-1d9c84df87a9   http://loinc.org          32207-3
#> 17  404daf3b-c444-4420-b62d-7ad250dc839f   http://loinc.org          32623-1
#> 18  b34721d1-9ba4-47ad-b6f6-092d08836808   http://loinc.org          72166-2
#> 19  419a2be0-6947-497f-a187-a60a29b3da50   http://loinc.org           8302-2
#> 20  df3d41e3-cde6-4ef6-8c52-4da8812d04a4   http://loinc.org          72514-3
#> 21  fdd0a486-5c96-4fca-baa8-12d138cb8c04   http://loinc.org          29463-7
#> 22  f6212802-2ac0-44a0-bb11-ef842784b386   http://loinc.org          39156-5
#> 23  fc0052b5-d384-4810-9050-3a7ac0438d8f   http://loinc.org          59576-9
#> 24  1b0076b9-dd15-46cf-8094-cfb8d09a7f5a   http://loinc.org          55284-4
#> 25  82022bc4-35b0-48d0-8bf5-b83443ee5510   http://loinc.org          72166-2
#> 26  28b3893b-74d4-4b4d-808e-a3b938aa3839   http://loinc.org           8302-2
#> 27  da50c832-6184-4428-9a07-840be794c5b5   http://loinc.org          72514-3
#> 28  9b24f946-e4ac-4bd5-9ad6-a62b9e46af87   http://loinc.org          29463-7
#> 29  137b7d5d-07b2-4561-b17d-de9d7ddc1a19   http://loinc.org          39156-5
#> 30  99f71a69-6728-410b-8f7e-888979d27888   http://loinc.org          59576-9
#> 31  563bfad3-9754-4db5-b99b-b3a53e3c08e7   http://loinc.org          55284-4
#> 32  da1be38e-d046-41d7-a37b-e6279ccddc73   http://loinc.org          72166-2
#> 33  7641669b-af7b-4c05-8399-86fb626f58bb   http://loinc.org           8302-2
#> 34  1fa00b20-c558-473d-b7c2-148f3b5f1674   http://loinc.org          72514-3
#> 35  8f3c5d96-b2ac-4aba-8f73-a98dd748afb1   http://loinc.org          29463-7
#> 36  aacb5d40-86e6-4862-8e21-d9c1fff7af3f   http://loinc.org          39156-5
#> 37  06409445-f75b-4e74-980e-a7e56966f6e1   http://loinc.org          59576-9
#> 38  a5747da9-af07-4ab1-ad2f-d4fac2fd8f43   http://loinc.org          55284-4
#> 39  e5bd19ee-dae5-4ee2-9ce7-ae9d5befdb35   http://loinc.org          72166-2
#> 40  7b1bb919-8887-46a3-a565-4959cdbaeacb   http://loinc.org           8302-2
#> 41  868daac1-8be0-4599-a3bd-749f8eb3ebaf   http://loinc.org          72514-3
#> 42  5eb46e83-5fef-4be1-812d-88d2e6943507   http://loinc.org          29463-7
#> 43  2518ddc1-1031-4933-a53f-2dc01ca4be03   http://loinc.org          39156-5
#> 44  bb212899-1b1c-4cf3-b522-3a48de2e6880   http://loinc.org          59576-9
#> 45  e98a1a52-6b4c-4412-8ef7-e2c92d2e0972   http://loinc.org          55284-4
#> 46  7f92011d-9e77-4036-9d6c-d2bed022131b   http://loinc.org          72166-2
#> 47  5e7ceb8e-714d-40b0-aa18-81c042bea44f   http://loinc.org           8302-2
#> 48  beb21f06-0baa-4d25-9d16-15600be15a93   http://loinc.org          72514-3
#> 49  f1cee0ae-4290-400d-a129-52ed8c389be9   http://loinc.org          29463-7
#> 50  d7e7ed01-8147-40a2-b41f-ebceb4c9aacf   http://loinc.org          39156-5
#> 51  594ace37-dc02-4ad7-a147-bf8e3e0535f2   http://loinc.org          59576-9
#> 52  cedfd069-2c83-4930-9675-879e6caaaa2a   http://loinc.org          55284-4
#> 53  486708e1-86e8-41f3-a3bc-4e15101ab2bd   http://loinc.org          72166-2
#> 54  9eeb47f9-9841-4d36-bd49-dc6fc10a78e1   http://loinc.org           8302-2
#> 55  a887a854-44db-4f61-8421-d8ebbc532077   http://loinc.org          72514-3
#> 56  8eb667b6-445a-4071-8a8b-ff648a51b06a   http://loinc.org          29463-7
#> 57  c3c54f91-988f-44cb-8d97-2b94cbb7a78a   http://loinc.org          39156-5
#> 58  b1e186ae-915e-433f-83ab-3d5cf70f843b   http://loinc.org          59576-9
#> 59  1155c32e-4489-4ee7-8452-27c48c3e03c5   http://loinc.org          55284-4
#> 60  4baab580-697d-4e91-92e9-0ee1f12fc695   http://loinc.org           6690-2
#> 61  5aa31e13-db5e-4296-bdc0-79ca8726e86e   http://loinc.org            789-8
#> 62  b3b27615-4a4d-4196-94d2-a2f611cc0462   http://loinc.org            718-7
#> 63  3cab3883-2393-4d8e-9fc3-bc12acefef6f   http://loinc.org           4544-3
#> 64  368ee1bd-2b50-4c45-aaf0-5643cdb914ff   http://loinc.org            787-2
#> 65  847219ff-0e28-4ef0-981d-cc27c9b5ce84   http://loinc.org            785-6
#> 66  f5589e7f-938c-4ed9-8fdf-f9305ff24a8e   http://loinc.org            786-4
#> 67  d108b1d5-0ee7-4225-b829-68d9642c005f   http://loinc.org          21000-5
#> 68  6507f695-524c-473c-98de-735c716660a3   http://loinc.org            777-3
#> 69  581c1059-cb28-4ee8-be0f-cbee0adb6082   http://loinc.org          32207-3
#> 70  43665ace-fe3b-4cd6-9c20-68a6acf14976   http://loinc.org          32623-1
#> 71  5020833a-85a3-4b50-9014-119bd074508b   http://loinc.org          72166-2
#> 72  c5cd3259-a1e0-42e4-8608-03f61ec6ea85   http://loinc.org           8302-2
#> 73  71538033-2b3a-4057-9f5a-97213238953b   http://loinc.org          72514-3
#> 74  588d5f31-0cff-4218-a5a9-163e7067512e   http://loinc.org          29463-7
#> 75  03cb40a4-b3b8-4809-9f2e-2a63e174e397   http://loinc.org          39156-5
#> 76  d36859e1-0302-40d7-8e66-f7525ccbc0f3   http://loinc.org          59576-9
#> 77  98fb7c36-ffcc-48eb-8886-f329739d7fba   http://loinc.org          55284-4
#> 78  ce429f8d-ca57-4a4a-baa5-8638136c32a8   http://loinc.org          72166-2
#> 79  ed5001ee-4da3-4d3b-8a31-86b102b1948b   http://loinc.org           8302-2
#> 80  8840fcdf-0baf-4fc5-95d3-645a93337488   http://loinc.org          72514-3
#> 81  ab2bf929-5c60-4143-a992-24c792be0b59   http://loinc.org          29463-7
#> 82  851abed2-8cfb-4867-8a0c-eb56e51bb990   http://loinc.org          39156-5
#> 83  305bbe04-d5c2-44ed-b19f-ca239fe12fb0   http://loinc.org          59576-9
#> 84  5715dcaa-0e26-4ed1-a680-3b48cab9a646   http://loinc.org          55284-4
#> 85  58d98770-e70a-4ad3-a671-8ebce593279f   http://loinc.org          72166-2
#> 86  23acfa61-91b0-4e47-9197-cc27e95ff53b   http://loinc.org           8302-2
#> 87  81861fd6-1e63-46a2-a568-7b5312f73b85   http://loinc.org          72514-3
#> 88  9833cf87-7597-47c1-8f04-efe8a0f0b2f0   http://loinc.org          29463-7
#> 89  77192094-49de-4306-8eaa-be9b1908e831   http://loinc.org          39156-5
#> 90  dc04e378-c403-4c38-8221-f755652a7c67   http://loinc.org          59576-9
#> 91  a8dd395b-3d65-45ce-8125-d0dcad7568d4   http://loinc.org          55284-4
#> 92  33c36791-d4c1-4872-bd09-738c7eaeb9a3   http://loinc.org          72166-2
#> 93  fe02b7c1-bd00-49e0-af8b-c5e3e2c79ce7   http://loinc.org           8302-2
#> 94  9b250f10-1eb8-400b-af42-aaaa7be40def   http://loinc.org          72514-3
#> 95  a1ddd866-7641-49d4-84c1-7ca0bb23ff9f   http://loinc.org          29463-7
#> 96  318a7d1e-1348-4076-9d14-123f931be1a5   http://loinc.org          39156-5
#> 97  2932c436-50cc-4ea1-8b37-c707f2f16aad   http://loinc.org          59576-9
#> 98  6c3f661d-ccc8-46b2-b80c-1c8839ac74a9   http://loinc.org          55284-4
#> 99  d976cf46-4c24-420e-a754-8a014ffe162a   http://loinc.org          72166-2
#> 100 0df9fcab-82dd-40b7-bf4d-c1090a4ee51b   http://loinc.org           8302-2
#> 101 eafbe6e1-76ed-42c7-b928-0383f916f330   http://loinc.org          72514-3
#> 102 5e45300f-6c17-4de7-90ef-559a6f197b77   http://loinc.org          29463-7
#> 103 18267523-5f6f-40df-a9c7-f08a4e07db7c   http://loinc.org          39156-5
#> 104 c37d696f-0952-4bbf-a4a8-eb8f615e3e58   http://loinc.org          59576-9
#> 105 075d7c87-13fd-47c6-b1a6-2b8685a2713d   http://loinc.org          55284-4
#> 106 088b77fe-a383-4c1f-807e-4a59624a5854   http://loinc.org          72166-2
#> 107 2644776f-787c-4645-ab8d-8f42131ad5bb   http://loinc.org           8302-2
#> 108 ba41dbdb-c2bf-438b-84aa-b0ec925ab0e2   http://loinc.org          72514-3
#> 109 442b8599-c917-4e19-ac1b-fc05e6233b88   http://loinc.org          29463-7
#> 110 6f5e9a4e-535c-4409-81d6-9bf71970024f   http://loinc.org          39156-5
#> 111 7ee7d2f1-1dce-450a-92c4-f8a1d88b61c3   http://loinc.org          59576-9
#> 112 12f14bc9-4721-4f82-bc1e-fff52ff2451e   http://loinc.org          55284-4
#> 113 c56a1c31-bf2f-4f2a-974d-6e8f9c207fbf   http://loinc.org          72166-2
#> 114 f4960bfa-4b15-4e46-af0d-488f31c33822   http://loinc.org          69453-9
#>                                                          code.coding.display
#> 1                                                                Body Height
#> 2              Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 3                                                                Body Weight
#> 4                                                            Body Mass Index
#> 5                      Body mass index (BMI) [Percentile] Per age and gender
#> 6                                                             Blood Pressure
#> 7                          Leukocytes [#/volume] in Blood by Automated count
#> 8                        Erythrocytes [#/volume] in Blood by Automated count
#> 9                                          Hemoglobin [Mass/volume] in Blood
#> 10                  Hematocrit [Volume Fraction] of Blood by Automated count
#> 11                                   MCV [Entitic volume] by Automated count
#> 12                                     MCH [Entitic mass] by Automated count
#> 13                                     MCHC [Mass/volume] by Automated count
#> 14        Erythrocyte distribution width [Entitic volume] by Automated count
#> 15                          Platelets [#/volume] in Blood by Automated count
#> 16  Platelet distribution width [Entitic volume] in Blood by Automated count
#> 17         Platelet mean volume [Entitic volume] in Blood by Automated count
#> 18                                               Tobacco smoking status NHIS
#> 19                                                               Body Height
#> 20             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 21                                                               Body Weight
#> 22                                                           Body Mass Index
#> 23                     Body mass index (BMI) [Percentile] Per age and gender
#> 24                                                            Blood Pressure
#> 25                                               Tobacco smoking status NHIS
#> 26                                                               Body Height
#> 27             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 28                                                               Body Weight
#> 29                                                           Body Mass Index
#> 30                     Body mass index (BMI) [Percentile] Per age and gender
#> 31                                                            Blood Pressure
#> 32                                               Tobacco smoking status NHIS
#> 33                                                               Body Height
#> 34             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 35                                                               Body Weight
#> 36                                                           Body Mass Index
#> 37                     Body mass index (BMI) [Percentile] Per age and gender
#> 38                                                            Blood Pressure
#> 39                                               Tobacco smoking status NHIS
#> 40                                                               Body Height
#> 41             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 42                                                               Body Weight
#> 43                                                           Body Mass Index
#> 44                     Body mass index (BMI) [Percentile] Per age and gender
#> 45                                                            Blood Pressure
#> 46                                               Tobacco smoking status NHIS
#> 47                                                               Body Height
#> 48             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 49                                                               Body Weight
#> 50                                                           Body Mass Index
#> 51                     Body mass index (BMI) [Percentile] Per age and gender
#> 52                                                            Blood Pressure
#> 53                                               Tobacco smoking status NHIS
#> 54                                                               Body Height
#> 55             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 56                                                               Body Weight
#> 57                                                           Body Mass Index
#> 58                     Body mass index (BMI) [Percentile] Per age and gender
#> 59                                                            Blood Pressure
#> 60                         Leukocytes [#/volume] in Blood by Automated count
#> 61                       Erythrocytes [#/volume] in Blood by Automated count
#> 62                                         Hemoglobin [Mass/volume] in Blood
#> 63                  Hematocrit [Volume Fraction] of Blood by Automated count
#> 64                                   MCV [Entitic volume] by Automated count
#> 65                                     MCH [Entitic mass] by Automated count
#> 66                                     MCHC [Mass/volume] by Automated count
#> 67        Erythrocyte distribution width [Entitic volume] by Automated count
#> 68                          Platelets [#/volume] in Blood by Automated count
#> 69  Platelet distribution width [Entitic volume] in Blood by Automated count
#> 70         Platelet mean volume [Entitic volume] in Blood by Automated count
#> 71                                               Tobacco smoking status NHIS
#> 72                                                               Body Height
#> 73             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 74                                                               Body Weight
#> 75                                                           Body Mass Index
#> 76                     Body mass index (BMI) [Percentile] Per age and gender
#> 77                                                            Blood Pressure
#> 78                                               Tobacco smoking status NHIS
#> 79                                                               Body Height
#> 80             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 81                                                               Body Weight
#> 82                                                           Body Mass Index
#> 83                     Body mass index (BMI) [Percentile] Per age and gender
#> 84                                                            Blood Pressure
#> 85                                               Tobacco smoking status NHIS
#> 86                                                               Body Height
#> 87             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 88                                                               Body Weight
#> 89                                                           Body Mass Index
#> 90                     Body mass index (BMI) [Percentile] Per age and gender
#> 91                                                            Blood Pressure
#> 92                                               Tobacco smoking status NHIS
#> 93                                                               Body Height
#> 94             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 95                                                               Body Weight
#> 96                                                           Body Mass Index
#> 97                     Body mass index (BMI) [Percentile] Per age and gender
#> 98                                                            Blood Pressure
#> 99                                               Tobacco smoking status NHIS
#> 100                                                              Body Height
#> 101            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 102                                                              Body Weight
#> 103                                                          Body Mass Index
#> 104                    Body mass index (BMI) [Percentile] Per age and gender
#> 105                                                           Blood Pressure
#> 106                                              Tobacco smoking status NHIS
#> 107                                                              Body Height
#> 108            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 109                                                              Body Weight
#> 110                                                          Body Mass Index
#> 111                    Body mass index (BMI) [Percentile] Per age and gender
#> 112                                                           Blood Pressure
#> 113                                              Tobacco smoking status NHIS
#> 114                        Cause of Death [US Standard Certificate of Death]