extract information from retained fields in Procedure component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Procedure component of FHIR Bundle, produce simple data.frame

process_Procedure(Procedure)

Arguments

Procedure

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Procedure(tbun$Procedure)
#>                                       id
#> 87  7f68ac33-b103-4fe9-b6da-f0e1556afa76
#> 135 c0dee9be-fe7f-4267-865c-8567116f2386
#> 190 0a9e0571-e059-4267-807c-ea020ec515f2
#> 202 8d440838-6e99-417a-a0a8-217fc573948c
#> 215 a8f29973-0d7a-4a1f-b572-65268457b38f
#> 246 4d300ced-3b75-4aee-bddf-ce1f0da73467
#> 253 64eccb84-b198-4f2b-a203-453bac382660
#> 266 f9aaf651-912d-4950-a571-890e25346745
#> 278 73f3655f-317a-402c-901a-f7d6321329a3
#>                                           subject    status
#> 87  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 135 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 190 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 202 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 215 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 246 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 253 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 266 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#> 278 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2 completed
#>         performedPeriod.start       performedPeriod.end            code.system
#> 87  1980-01-07T06:25:09-05:00 1980-01-07T06:40:09-05:00 http://snomed.info/sct
#> 135 1981-07-01T07:25:09-04:00 1981-07-01T07:40:09-04:00 http://snomed.info/sct
#> 190 1982-06-16T07:25:09-04:00 1982-06-16T07:40:09-04:00 http://snomed.info/sct
#> 202 1982-07-07T07:25:09-04:00 1982-07-07T07:40:09-04:00 http://snomed.info/sct
#> 215 1983-03-24T06:25:09-05:00 1983-03-24T06:40:09-05:00 http://snomed.info/sct
#> 246 1983-08-31T07:25:09-04:00 1983-08-31T07:40:09-04:00 http://snomed.info/sct
#> 253 1984-06-19T07:25:09-04:00 1984-06-19T10:10:09-04:00 http://snomed.info/sct
#> 266 1984-07-18T07:25:09-04:00 1984-07-18T07:40:09-04:00 http://snomed.info/sct
#> 278 1985-07-24T07:25:09-04:00 1985-07-24T07:40:09-04:00 http://snomed.info/sct
#>     code.code                                             code.display
#> 87  313191000                                  Injection of adrenaline
#> 135 430193006                    Medication Reconciliation (procedure)
#> 190 430193006                    Medication Reconciliation (procedure)
#> 202 430193006                    Medication Reconciliation (procedure)
#> 215 313191000                                  Injection of adrenaline
#> 246 430193006                    Medication Reconciliation (procedure)
#> 253  54550000 Seizure Count Cerebral Cortex Electroencephalogram (EEG)
#> 266 430193006                    Medication Reconciliation (procedure)
#> 278 430193006                    Medication Reconciliation (procedure)