extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame

extract information from retained fields in Observation component of FHIR Bundle, produce simple data.frame

process_Observation(Observation)

Arguments

Observation

component of FHIR.bundle instance

Value

data.frame

Examples

testf = system.file("json/Vince741_Rogahn59_6fa3d4ab-c0b6-424a-89d8-7d9105129296.json",
   package="BiocFHIR")
tbun = process_fhir_bundle(testf)
process_Observation(tbun$Observation)
#>                                       id
#> 33  d8d29164-6f51-411f-9ac3-a8fd4746bbb4
#> 34  0cb0cc06-d732-45cf-b79a-29127915baa2
#> 35  e16a09b0-b01f-44cb-b301-d606729e4b38
#> 36  9625b98f-bc58-453c-8664-f1fe185840a8
#> 37  690c8151-8c6b-4c1a-81d9-1071818592cc
#> 38  a91b50ea-2596-4f2c-9a3c-4d38054a89eb
#> 39  74b16bae-d8b1-4cf9-9281-ad3736348641
#> 40  911c76fe-c953-4bf2-ba74-d71b44399296
#> 41  013b4774-cea5-437b-a6ea-1e0e0dcde8d9
#> 42  895ccd65-5bec-4963-9bfd-824865121280
#> 43  0d72572d-2c30-4d7d-8aa2-14475e2f3962
#> 44  810cc728-d102-44c1-9ea0-6b08f2d0edc5
#> 45  899058cb-26ec-4be6-bd1c-43cac0356d54
#> 46  9ef4647c-d144-4195-a8c4-89ab15910d1b
#> 47  b4b17f38-8920-4634-a9ce-b0d2e7833bcc
#> 48  18b51ef9-e7d7-4656-95f8-1d9c84df87a9
#> 49  404daf3b-c444-4420-b62d-7ad250dc839f
#> 50  b34721d1-9ba4-47ad-b6f6-092d08836808
#> 56  419a2be0-6947-497f-a187-a60a29b3da50
#> 57  df3d41e3-cde6-4ef6-8c52-4da8812d04a4
#> 58  fdd0a486-5c96-4fca-baa8-12d138cb8c04
#> 59  f6212802-2ac0-44a0-bb11-ef842784b386
#> 60  fc0052b5-d384-4810-9050-3a7ac0438d8f
#> 61  1b0076b9-dd15-46cf-8094-cfb8d09a7f5a
#> 62  82022bc4-35b0-48d0-8bf5-b83443ee5510
#> 67  28b3893b-74d4-4b4d-808e-a3b938aa3839
#> 68  da50c832-6184-4428-9a07-840be794c5b5
#> 69  9b24f946-e4ac-4bd5-9ad6-a62b9e46af87
#> 70  137b7d5d-07b2-4561-b17d-de9d7ddc1a19
#> 71  99f71a69-6728-410b-8f7e-888979d27888
#> 72  563bfad3-9754-4db5-b99b-b3a53e3c08e7
#> 73  da1be38e-d046-41d7-a37b-e6279ccddc73
#> 96  7641669b-af7b-4c05-8399-86fb626f58bb
#> 97  1fa00b20-c558-473d-b7c2-148f3b5f1674
#> 98  8f3c5d96-b2ac-4aba-8f73-a98dd748afb1
#> 99  aacb5d40-86e6-4862-8e21-d9c1fff7af3f
#> 100 06409445-f75b-4e74-980e-a7e56966f6e1
#> 101 a5747da9-af07-4ab1-ad2f-d4fac2fd8f43
#> 102 e5bd19ee-dae5-4ee2-9ce7-ae9d5befdb35
#> 128 7b1bb919-8887-46a3-a565-4959cdbaeacb
#> 129 868daac1-8be0-4599-a3bd-749f8eb3ebaf
#> 130 5eb46e83-5fef-4be1-812d-88d2e6943507
#> 131 2518ddc1-1031-4933-a53f-2dc01ca4be03
#> 132 bb212899-1b1c-4cf3-b522-3a48de2e6880
#> 133 e98a1a52-6b4c-4412-8ef7-e2c92d2e0972
#> 134 7f92011d-9e77-4036-9d6c-d2bed022131b
#> 142 5e7ceb8e-714d-40b0-aa18-81c042bea44f
#> 143 beb21f06-0baa-4d25-9d16-15600be15a93
#> 144 f1cee0ae-4290-400d-a129-52ed8c389be9
#> 145 d7e7ed01-8147-40a2-b41f-ebceb4c9aacf
#> 146 594ace37-dc02-4ad7-a147-bf8e3e0535f2
#> 147 cedfd069-2c83-4930-9675-879e6caaaa2a
#> 148 486708e1-86e8-41f3-a3bc-4e15101ab2bd
#> 172 9eeb47f9-9841-4d36-bd49-dc6fc10a78e1
#> 173 a887a854-44db-4f61-8421-d8ebbc532077
#> 174 8eb667b6-445a-4071-8a8b-ff648a51b06a
#> 175 c3c54f91-988f-44cb-8d97-2b94cbb7a78a
#> 176 b1e186ae-915e-433f-83ab-3d5cf70f843b
#> 177 1155c32e-4489-4ee7-8452-27c48c3e03c5
#> 178 4baab580-697d-4e91-92e9-0ee1f12fc695
#> 179 5aa31e13-db5e-4296-bdc0-79ca8726e86e
#> 180 b3b27615-4a4d-4196-94d2-a2f611cc0462
#> 181 3cab3883-2393-4d8e-9fc3-bc12acefef6f
#> 182 368ee1bd-2b50-4c45-aaf0-5643cdb914ff
#> 183 847219ff-0e28-4ef0-981d-cc27c9b5ce84
#> 184 f5589e7f-938c-4ed9-8fdf-f9305ff24a8e
#> 185 d108b1d5-0ee7-4225-b829-68d9642c005f
#> 186 6507f695-524c-473c-98de-735c716660a3
#> 187 581c1059-cb28-4ee8-be0f-cbee0adb6082
#> 188 43665ace-fe3b-4cd6-9c20-68a6acf14976
#> 189 5020833a-85a3-4b50-9014-119bd074508b
#> 195 c5cd3259-a1e0-42e4-8608-03f61ec6ea85
#> 196 71538033-2b3a-4057-9f5a-97213238953b
#> 197 588d5f31-0cff-4218-a5a9-163e7067512e
#> 198 03cb40a4-b3b8-4809-9f2e-2a63e174e397
#> 199 d36859e1-0302-40d7-8e66-f7525ccbc0f3
#> 200 98fb7c36-ffcc-48eb-8886-f329739d7fba
#> 201 ce429f8d-ca57-4a4a-baa5-8638136c32a8
#> 224 ed5001ee-4da3-4d3b-8a31-86b102b1948b
#> 225 8840fcdf-0baf-4fc5-95d3-645a93337488
#> 226 ab2bf929-5c60-4143-a992-24c792be0b59
#> 227 851abed2-8cfb-4867-8a0c-eb56e51bb990
#> 228 305bbe04-d5c2-44ed-b19f-ca239fe12fb0
#> 229 5715dcaa-0e26-4ed1-a680-3b48cab9a646
#> 230 58d98770-e70a-4ad3-a671-8ebce593279f
#> 239 23acfa61-91b0-4e47-9197-cc27e95ff53b
#> 240 81861fd6-1e63-46a2-a568-7b5312f73b85
#> 241 9833cf87-7597-47c1-8f04-efe8a0f0b2f0
#> 242 77192094-49de-4306-8eaa-be9b1908e831
#> 243 dc04e378-c403-4c38-8221-f755652a7c67
#> 244 a8dd395b-3d65-45ce-8125-d0dcad7568d4
#> 245 33c36791-d4c1-4872-bd09-738c7eaeb9a3
#> 259 fe02b7c1-bd00-49e0-af8b-c5e3e2c79ce7
#> 260 9b250f10-1eb8-400b-af42-aaaa7be40def
#> 261 a1ddd866-7641-49d4-84c1-7ca0bb23ff9f
#> 262 318a7d1e-1348-4076-9d14-123f931be1a5
#> 263 2932c436-50cc-4ea1-8b37-c707f2f16aad
#> 264 6c3f661d-ccc8-46b2-b80c-1c8839ac74a9
#> 265 d976cf46-4c24-420e-a754-8a014ffe162a
#> 271 0df9fcab-82dd-40b7-bf4d-c1090a4ee51b
#> 272 eafbe6e1-76ed-42c7-b928-0383f916f330
#> 273 5e45300f-6c17-4de7-90ef-559a6f197b77
#> 274 18267523-5f6f-40df-a9c7-f08a4e07db7c
#> 275 c37d696f-0952-4bbf-a4a8-eb8f615e3e58
#> 276 075d7c87-13fd-47c6-b1a6-2b8685a2713d
#> 277 088b77fe-a383-4c1f-807e-4a59624a5854
#> 285 2644776f-787c-4645-ab8d-8f42131ad5bb
#> 286 ba41dbdb-c2bf-438b-84aa-b0ec925ab0e2
#> 287 442b8599-c917-4e19-ac1b-fc05e6233b88
#> 288 6f5e9a4e-535c-4409-81d6-9bf71970024f
#> 289 7ee7d2f1-1dce-450a-92c4-f8a1d88b61c3
#> 290 12f14bc9-4721-4f82-bc1e-fff52ff2451e
#> 291 c56a1c31-bf2f-4f2a-974d-6e8f9c207fbf
#> 298 f4960bfa-4b15-4e46-af0d-488f31c33822
#>                                 subject.reference code.coding.system
#> 33  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 34  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 35  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 36  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 37  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 38  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 39  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 40  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 41  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 42  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 43  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 44  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 45  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 46  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 47  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 48  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 49  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 50  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 56  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 57  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 58  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 59  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 60  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 61  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 62  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 67  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 68  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 69  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 70  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 71  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 72  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 73  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 96  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 97  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 98  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 99  urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 100 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 101 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 102 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 128 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 129 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 130 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 131 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 132 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 133 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 134 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 142 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 143 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 144 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 145 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 146 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 147 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 148 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 172 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 173 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 174 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 175 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 176 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 177 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 178 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 179 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 180 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 181 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 182 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 183 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 184 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 185 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 186 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 187 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 188 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 189 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 195 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 196 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 197 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 198 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 199 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 200 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 201 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 224 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 225 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 226 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 227 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 228 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 229 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 230 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 239 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 240 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 241 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 242 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 243 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 244 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 245 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 259 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 260 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 261 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 262 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 263 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 264 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 265 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 271 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 272 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 273 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 274 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 275 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 276 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 277 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 285 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 286 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 287 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 288 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 289 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 290 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 291 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#> 298 urn:uuid:eedf9986-9cf8-4e90-bf68-12a6dd9a31c2   http://loinc.org
#>     code.coding.code
#> 33            8302-2
#> 34           72514-3
#> 35           29463-7
#> 36           39156-5
#> 37           59576-9
#> 38           55284-4
#> 39            6690-2
#> 40             789-8
#> 41             718-7
#> 42            4544-3
#> 43             787-2
#> 44             785-6
#> 45             786-4
#> 46           21000-5
#> 47             777-3
#> 48           32207-3
#> 49           32623-1
#> 50           72166-2
#> 56            8302-2
#> 57           72514-3
#> 58           29463-7
#> 59           39156-5
#> 60           59576-9
#> 61           55284-4
#> 62           72166-2
#> 67            8302-2
#> 68           72514-3
#> 69           29463-7
#> 70           39156-5
#> 71           59576-9
#> 72           55284-4
#> 73           72166-2
#> 96            8302-2
#> 97           72514-3
#> 98           29463-7
#> 99           39156-5
#> 100          59576-9
#> 101          55284-4
#> 102          72166-2
#> 128           8302-2
#> 129          72514-3
#> 130          29463-7
#> 131          39156-5
#> 132          59576-9
#> 133          55284-4
#> 134          72166-2
#> 142           8302-2
#> 143          72514-3
#> 144          29463-7
#> 145          39156-5
#> 146          59576-9
#> 147          55284-4
#> 148          72166-2
#> 172           8302-2
#> 173          72514-3
#> 174          29463-7
#> 175          39156-5
#> 176          59576-9
#> 177          55284-4
#> 178           6690-2
#> 179            789-8
#> 180            718-7
#> 181           4544-3
#> 182            787-2
#> 183            785-6
#> 184            786-4
#> 185          21000-5
#> 186            777-3
#> 187          32207-3
#> 188          32623-1
#> 189          72166-2
#> 195           8302-2
#> 196          72514-3
#> 197          29463-7
#> 198          39156-5
#> 199          59576-9
#> 200          55284-4
#> 201          72166-2
#> 224           8302-2
#> 225          72514-3
#> 226          29463-7
#> 227          39156-5
#> 228          59576-9
#> 229          55284-4
#> 230          72166-2
#> 239           8302-2
#> 240          72514-3
#> 241          29463-7
#> 242          39156-5
#> 243          59576-9
#> 244          55284-4
#> 245          72166-2
#> 259           8302-2
#> 260          72514-3
#> 261          29463-7
#> 262          39156-5
#> 263          59576-9
#> 264          55284-4
#> 265          72166-2
#> 271           8302-2
#> 272          72514-3
#> 273          29463-7
#> 274          39156-5
#> 275          59576-9
#> 276          55284-4
#> 277          72166-2
#> 285           8302-2
#> 286          72514-3
#> 287          29463-7
#> 288          39156-5
#> 289          59576-9
#> 290          55284-4
#> 291          72166-2
#> 298          69453-9
#>                                                          code.coding.display
#> 33                                                               Body Height
#> 34             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 35                                                               Body Weight
#> 36                                                           Body Mass Index
#> 37                     Body mass index (BMI) [Percentile] Per age and gender
#> 38                                                            Blood Pressure
#> 39                         Leukocytes [#/volume] in Blood by Automated count
#> 40                       Erythrocytes [#/volume] in Blood by Automated count
#> 41                                         Hemoglobin [Mass/volume] in Blood
#> 42                  Hematocrit [Volume Fraction] of Blood by Automated count
#> 43                                   MCV [Entitic volume] by Automated count
#> 44                                     MCH [Entitic mass] by Automated count
#> 45                                     MCHC [Mass/volume] by Automated count
#> 46        Erythrocyte distribution width [Entitic volume] by Automated count
#> 47                          Platelets [#/volume] in Blood by Automated count
#> 48  Platelet distribution width [Entitic volume] in Blood by Automated count
#> 49         Platelet mean volume [Entitic volume] in Blood by Automated count
#> 50                                               Tobacco smoking status NHIS
#> 56                                                               Body Height
#> 57             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 58                                                               Body Weight
#> 59                                                           Body Mass Index
#> 60                     Body mass index (BMI) [Percentile] Per age and gender
#> 61                                                            Blood Pressure
#> 62                                               Tobacco smoking status NHIS
#> 67                                                               Body Height
#> 68             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 69                                                               Body Weight
#> 70                                                           Body Mass Index
#> 71                     Body mass index (BMI) [Percentile] Per age and gender
#> 72                                                            Blood Pressure
#> 73                                               Tobacco smoking status NHIS
#> 96                                                               Body Height
#> 97             Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 98                                                               Body Weight
#> 99                                                           Body Mass Index
#> 100                    Body mass index (BMI) [Percentile] Per age and gender
#> 101                                                           Blood Pressure
#> 102                                              Tobacco smoking status NHIS
#> 128                                                              Body Height
#> 129            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 130                                                              Body Weight
#> 131                                                          Body Mass Index
#> 132                    Body mass index (BMI) [Percentile] Per age and gender
#> 133                                                           Blood Pressure
#> 134                                              Tobacco smoking status NHIS
#> 142                                                              Body Height
#> 143            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 144                                                              Body Weight
#> 145                                                          Body Mass Index
#> 146                    Body mass index (BMI) [Percentile] Per age and gender
#> 147                                                           Blood Pressure
#> 148                                              Tobacco smoking status NHIS
#> 172                                                              Body Height
#> 173            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 174                                                              Body Weight
#> 175                                                          Body Mass Index
#> 176                    Body mass index (BMI) [Percentile] Per age and gender
#> 177                                                           Blood Pressure
#> 178                        Leukocytes [#/volume] in Blood by Automated count
#> 179                      Erythrocytes [#/volume] in Blood by Automated count
#> 180                                        Hemoglobin [Mass/volume] in Blood
#> 181                 Hematocrit [Volume Fraction] of Blood by Automated count
#> 182                                  MCV [Entitic volume] by Automated count
#> 183                                    MCH [Entitic mass] by Automated count
#> 184                                    MCHC [Mass/volume] by Automated count
#> 185       Erythrocyte distribution width [Entitic volume] by Automated count
#> 186                         Platelets [#/volume] in Blood by Automated count
#> 187 Platelet distribution width [Entitic volume] in Blood by Automated count
#> 188        Platelet mean volume [Entitic volume] in Blood by Automated count
#> 189                                              Tobacco smoking status NHIS
#> 195                                                              Body Height
#> 196            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 197                                                              Body Weight
#> 198                                                          Body Mass Index
#> 199                    Body mass index (BMI) [Percentile] Per age and gender
#> 200                                                           Blood Pressure
#> 201                                              Tobacco smoking status NHIS
#> 224                                                              Body Height
#> 225            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 226                                                              Body Weight
#> 227                                                          Body Mass Index
#> 228                    Body mass index (BMI) [Percentile] Per age and gender
#> 229                                                           Blood Pressure
#> 230                                              Tobacco smoking status NHIS
#> 239                                                              Body Height
#> 240            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 241                                                              Body Weight
#> 242                                                          Body Mass Index
#> 243                    Body mass index (BMI) [Percentile] Per age and gender
#> 244                                                           Blood Pressure
#> 245                                              Tobacco smoking status NHIS
#> 259                                                              Body Height
#> 260            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 261                                                              Body Weight
#> 262                                                          Body Mass Index
#> 263                    Body mass index (BMI) [Percentile] Per age and gender
#> 264                                                           Blood Pressure
#> 265                                              Tobacco smoking status NHIS
#> 271                                                              Body Height
#> 272            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 273                                                              Body Weight
#> 274                                                          Body Mass Index
#> 275                    Body mass index (BMI) [Percentile] Per age and gender
#> 276                                                           Blood Pressure
#> 277                                              Tobacco smoking status NHIS
#> 285                                                              Body Height
#> 286            Pain severity - 0-10 verbal numeric rating [Score] - Reported
#> 287                                                              Body Weight
#> 288                                                          Body Mass Index
#> 289                    Body mass index (BMI) [Percentile] Per age and gender
#> 290                                                           Blood Pressure
#> 291                                              Tobacco smoking status NHIS
#> 298                        Cause of Death [US Standard Certificate of Death]
#>     valueQuantity.value valueQuantity.unit      valueQuantity.system
#> 33          115.1299086                 cm http://unitsofmeasure.org
#> 34            0.5095500            {score} http://unitsofmeasure.org
#> 35           22.1132051                 kg http://unitsofmeasure.org
#> 36           16.6830470              kg/m2 http://unitsofmeasure.org
#> 37           78.5156581                  % http://unitsofmeasure.org
#> 38                   NA               <NA>                      <NA>
#> 39            8.6842718            10*3/uL http://unitsofmeasure.org
#> 40            4.4991114            10*6/uL http://unitsofmeasure.org
#> 41           12.5191333               g/dL http://unitsofmeasure.org
#> 42           42.3043307                  % http://unitsofmeasure.org
#> 43           86.5492304                 fL http://unitsofmeasure.org
#> 44           27.2504072                 pg http://unitsofmeasure.org
#> 45           34.8337997               g/dL http://unitsofmeasure.org
#> 46           45.5063513                 fL http://unitsofmeasure.org
#> 47          442.7691329            10*3/uL http://unitsofmeasure.org
#> 48          314.7454650                 fL http://unitsofmeasure.org
#> 49           11.3161486                 fL http://unitsofmeasure.org
#> 50                   NA               <NA>                      <NA>
#> 56          121.5137382                 cm http://unitsofmeasure.org
#> 57            2.0334647            {score} http://unitsofmeasure.org
#> 58           25.3852385                 kg http://unitsofmeasure.org
#> 59           17.1921616              kg/m2 http://unitsofmeasure.org
#> 60           80.8867927                  % http://unitsofmeasure.org
#> 61                   NA               <NA>                      <NA>
#> 62                   NA               <NA>                      <NA>
#> 67          127.6389989                 cm http://unitsofmeasure.org
#> 68            0.7826692            {score} http://unitsofmeasure.org
#> 69           28.9185156                 kg http://unitsofmeasure.org
#> 70           17.7504439              kg/m2 http://unitsofmeasure.org
#> 71           81.7353053                  % http://unitsofmeasure.org
#> 72                   NA               <NA>                      <NA>
#> 73                   NA               <NA>                      <NA>
#> 96          133.7419595                 cm http://unitsofmeasure.org
#> 97            0.3494132            {score} http://unitsofmeasure.org
#> 98           32.7486569                 kg http://unitsofmeasure.org
#> 99           18.3087262              kg/m2 http://unitsofmeasure.org
#> 100          81.2050558                  % http://unitsofmeasure.org
#> 101                  NA               <NA>                      <NA>
#> 102                  NA               <NA>                      <NA>
#> 128         138.8314139                 cm http://unitsofmeasure.org
#> 129           3.4349551            {score} http://unitsofmeasure.org
#> 130          36.3442741                 kg http://unitsofmeasure.org
#> 131          18.8564748              kg/m2 http://unitsofmeasure.org
#> 132          79.9408201                  % http://unitsofmeasure.org
#> 133                  NA               <NA>                      <NA>
#> 134                  NA               <NA>                      <NA>
#> 142         138.8314139                 cm http://unitsofmeasure.org
#> 143           2.7158927            {score} http://unitsofmeasure.org
#> 144          36.3442741                 kg http://unitsofmeasure.org
#> 145          18.8564748              kg/m2 http://unitsofmeasure.org
#> 146          79.9408201                  % http://unitsofmeasure.org
#> 147                  NA               <NA>                      <NA>
#> 148                  NA               <NA>                      <NA>
#> 172         143.3191983                 cm http://unitsofmeasure.org
#> 173           0.6674318            {score} http://unitsofmeasure.org
#> 174          40.7239421                 kg http://unitsofmeasure.org
#> 175          19.8262725              kg/m2 http://unitsofmeasure.org
#> 176          83.0438533                  % http://unitsofmeasure.org
#> 177                  NA               <NA>                      <NA>
#> 178           7.0617650            10*3/uL http://unitsofmeasure.org
#> 179           5.2542715            10*6/uL http://unitsofmeasure.org
#> 180          13.9524377               g/dL http://unitsofmeasure.org
#> 181          35.9177867                  % http://unitsofmeasure.org
#> 182          91.5592746                 fL http://unitsofmeasure.org
#> 183          29.4697459                 pg http://unitsofmeasure.org
#> 184          35.6605989               g/dL http://unitsofmeasure.org
#> 185          43.7514134                 fL http://unitsofmeasure.org
#> 186         171.6531336            10*3/uL http://unitsofmeasure.org
#> 187         181.9547751                 fL http://unitsofmeasure.org
#> 188           9.4334416                 fL http://unitsofmeasure.org
#> 189                  NA               <NA>                      <NA>
#> 195         143.7388827                 cm http://unitsofmeasure.org
#> 196           1.1719911            {score} http://unitsofmeasure.org
#> 197          41.1477289                 kg http://unitsofmeasure.org
#> 198          19.9157809              kg/m2 http://unitsofmeasure.org
#> 199          83.2022415                  % http://unitsofmeasure.org
#> 200                  NA               <NA>                      <NA>
#> 201                  NA               <NA>                      <NA>
#> 224         149.3178474                 cm http://unitsofmeasure.org
#> 225           2.8462028            {score} http://unitsofmeasure.org
#> 226          46.5422889                 kg http://unitsofmeasure.org
#> 227          20.8748949              kg/m2 http://unitsofmeasure.org
#> 228          84.1353202                  % http://unitsofmeasure.org
#> 229                  NA               <NA>                      <NA>
#> 230                  NA               <NA>                      <NA>
#> 239         150.3875442                 cm http://unitsofmeasure.org
#> 240           3.4065984            {score} http://unitsofmeasure.org
#> 241          47.4892903                 kg http://unitsofmeasure.org
#> 242          20.9977105              kg/m2 http://unitsofmeasure.org
#> 243          84.0509261                  % http://unitsofmeasure.org
#> 244                  NA               <NA>                      <NA>
#> 245                  NA               <NA>                      <NA>
#> 259         156.4194178                 cm http://unitsofmeasure.org
#> 260           1.1908458            {score} http://unitsofmeasure.org
#> 261          52.3575057                 kg http://unitsofmeasure.org
#> 262          21.3992040              kg/m2 http://unitsofmeasure.org
#> 263          82.3427755                  % http://unitsofmeasure.org
#> 264                  NA               <NA>                      <NA>
#> 265                  NA               <NA>                      <NA>
#> 271         164.7449393                 cm http://unitsofmeasure.org
#> 272           1.4464437            {score} http://unitsofmeasure.org
#> 273          58.7573699                 kg http://unitsofmeasure.org
#> 274          21.6490170              kg/m2 http://unitsofmeasure.org
#> 275          78.0663162                  % http://unitsofmeasure.org
#> 276                  NA               <NA>                      <NA>
#> 277                  NA               <NA>                      <NA>
#> 285         170.5288757                 cm http://unitsofmeasure.org
#> 286           2.6526610            {score} http://unitsofmeasure.org
#> 287          64.2981597                 kg http://unitsofmeasure.org
#> 288          22.1107099              kg/m2 http://unitsofmeasure.org
#> 289          75.6173772                  % http://unitsofmeasure.org
#> 290                  NA               <NA>                      <NA>
#> 291                  NA               <NA>                      <NA>
#> 298                  NA               <NA>                      <NA>
#>     valueQuantity.code         effectiveDateTime                        issued
#> 33                  cm 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 34             {score} 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 35                  kg 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 36               kg/m2 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 37                   % 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 38                <NA> 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 39             10*3/uL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 40             10*6/uL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 41                g/dL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 42                   % 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 43                  fL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 44                  pg 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 45                g/dL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 46                  fL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 47             10*3/uL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 48                  fL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 49                  fL 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 50                <NA> 1977-06-08T07:25:09-04:00 1977-06-08T07:25:09.496-04:00
#> 56                  cm 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 57             {score} 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 58                  kg 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 59               kg/m2 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 60                   % 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 61                <NA> 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 62                <NA> 1978-06-14T07:25:09-04:00 1978-06-14T07:25:09.496-04:00
#> 67                  cm 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 68             {score} 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 69                  kg 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 70               kg/m2 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 71                   % 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 72                <NA> 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 73                <NA> 1979-06-20T07:25:09-04:00 1979-06-20T07:25:09.496-04:00
#> 96                  cm 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 97             {score} 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 98                  kg 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 99               kg/m2 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 100                  % 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 101               <NA> 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 102               <NA> 1980-06-25T07:25:09-04:00 1980-06-25T07:25:09.496-04:00
#> 128                 cm 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 129            {score} 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 130                 kg 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 131              kg/m2 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 132                  % 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 133               <NA> 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 134               <NA> 1981-07-01T07:25:09-04:00 1981-07-01T07:25:09.496-04:00
#> 142                 cm 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 143            {score} 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 144                 kg 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 145              kg/m2 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 146                  % 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 147               <NA> 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 148               <NA> 1981-07-15T07:25:09-04:00 1981-07-15T07:25:09.496-04:00
#> 172                 cm 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 173            {score} 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 174                 kg 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 175              kg/m2 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 176                  % 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 177               <NA> 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 178            10*3/uL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 179            10*6/uL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 180               g/dL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 181                  % 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 182                 fL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 183                 pg 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 184               g/dL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 185                 fL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 186            10*3/uL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 187                 fL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 188                 fL 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 189               <NA> 1982-06-16T07:25:09-04:00 1982-06-16T07:25:09.496-04:00
#> 195                 cm 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 196            {score} 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 197                 kg 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 198              kg/m2 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 199                  % 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 200               <NA> 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 201               <NA> 1982-07-07T07:25:09-04:00 1982-07-07T07:25:09.496-04:00
#> 224                 cm 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 225            {score} 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 226                 kg 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 227              kg/m2 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 228                  % 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 229               <NA> 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 230               <NA> 1983-07-13T07:25:09-04:00 1983-07-13T07:25:09.496-04:00
#> 239                 cm 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 240            {score} 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 241                 kg 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 242              kg/m2 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 243                  % 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 244               <NA> 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 245               <NA> 1983-08-31T07:25:09-04:00 1983-08-31T07:25:09.496-04:00
#> 259                 cm 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 260            {score} 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 261                 kg 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 262              kg/m2 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 263                  % 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 264               <NA> 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 265               <NA> 1984-07-18T07:25:09-04:00 1984-07-18T07:25:09.496-04:00
#> 271                 cm 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 272            {score} 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 273                 kg 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 274              kg/m2 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 275                  % 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 276               <NA> 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 277               <NA> 1985-07-24T07:25:09-04:00 1985-07-24T07:25:09.496-04:00
#> 285                 cm 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 286            {score} 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 287                 kg 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 288              kg/m2 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 289                  % 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 290               <NA> 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 291               <NA> 1986-07-30T07:25:09-04:00 1986-07-30T07:25:09.496-04:00
#> 298               <NA> 1987-05-20T07:25:09-04:00 1987-05-20T07:25:09.496-04:00